+++ src/store/index.ts
@@ -1,24 +1,8 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
+import appModule from './appModule'
 
 Vue.use(Vuex)
-
-interface appModuleState {
-  navTitle: string
-}
-
-const appModule = {
-  namespaced: true,
-  state: () => ({
-    navTitle: ''
-  }),
-  mutations: {
-    setNavTitle: (state: appModuleState, payload: { title: string }) => {
-      state.navTitle = payload.title
-    }
-  },
-  actions: {}
-}
 
 export default new Vuex.Store({
   modules: {
